СТРАДАНИЕ_1: Я несколько часов пытался привести бд к читаемому виду,я изменил кучу файлов где нужно было поменять значения на Long,я хотел сделать все аккуратно в итоге я сделал больше ошибок нежели исправлений, время 23:55 я кодил с БЕЗ ШУТОК с 5:50(с учетом перерывов на бутер и тд) но я очень устал и продолжу уже завтра

+++ src/main/java/kg/attractor/jobsearch_remake/service/ContactInfoService.java
@@ -32,7 +32,6 @@ public class ContactInfoService {
         dtos.forEach(dto -> contactInfoRepository.save(toModel(resumeId, dto)));
     }
 
-
     public void updateForResume(Integer resumeId, List<ContactInfoDto> dtos) {
         if (dtos == null) return;
         log.info("Обновление контактов для резюме id: {}", resumeId);
@@ -47,7 +46,7 @@ public class ContactInfoService {
 
     private ContactInfoDto toDto(ContactInfo info) {
         return ContactInfoDto.builder()
-                .id(info.getId())
+                .id(info.getId().intValue()) // ✅ Long → Integer
                 .resumeId(info.getResumeId())
                 .typeId(info.getTypeId())
                 .value(info.getValue())
